refactor(#217): flatten files structure
