+++ src/app/Background/TTS/TTSManager/index.ts
@@ -2,7 +2,7 @@ import { GoogleTTS } from '@translate-tools/core/tts/GoogleTTS';
 import { LingvaTTS } from '@translate-tools/core/tts/LingvaTTS';
 import { TTSProvider } from '@translate-tools/core/tts';
 
-import { SerializedSpeaker, TTSKey, TTSStorage } from './TTSStorage';
+import { SerializedSpeaker, TTSKey, TTSStorage } from '../TTSStorage';
 import { tryLoadTTSCode } from './utils';
 
 export type CustomTTS = SerializedSpeaker & {
