+++ src/app/Background/TTS/TTSManager/utils.ts
